requirements.txt
starlette==0.14.2
transformers==4.9.2
uvicorn==0.15.0
fastapi==0.68.0
python-dotenv==0.19.0
